% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaboliteqc.R
\name{byboxplot}
\alias{byboxplot}
\title{Draw boxplots for subsets of columns}
\usage{
byboxplot(mat, by, ylab, ylim = NULL)
}
\arguments{
\item{mat}{Matrix}

\item{by}{Factor of length \code{ncol(mat)} used to group columns
into subsets}

\item{ylab}{Label for y-axis}

\item{ylim}{Range of y-axis}
}
\value{
Returns \code{NULL} invisibly.
}
\description{
Draw boxplots for subsets of columns
}

