% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaboliteqc.R
\name{bybyboxplot}
\alias{bybyboxplot}
\title{Draw boxplots for rectangular subsets of matrix}
\usage{
bybyboxplot(mat, by.x, by.y, main = NULL, ylim = NULL)
}
\arguments{
\item{mat}{Matrix}

\item{by.x}{Factor to group columns of \code{mat} into subsets}

\item{by.y}{Factor to group rows of \code{mat} into subsets}

\item{main}{Title}

\item{ylim}{y-axis range to use for all subplots}
}
\value{
Returns \code{NULL} invisibly.
}
\description{
Draw boxplots for rectangular subsets of matrix
}
\details{
The \code{by.x} argument has suffix "x" because the
    subsets of columns that \code{by.x} creates will be drawn along
    the x-axis.  Similarly, the \code{by.y} argument has suffix "y"
    because the subsets of rows that \code{by.y} creates will be
    drawn along the y-axis.
}

