% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaboliteqc.R
\name{columnwise}
\alias{columnwise}
\title{Apply a function columnwise to groups of rows}
\usage{
columnwise(f, mat, by, ...)
}
\arguments{
\item{f}{Function}

\item{mat}{Matrix}

\item{by}{Factor of length \code{nrow(mat)} used to group rows}

\item{\dots}{Arguments passed on to \code{f}}
}
\value{
A matrix with \code{ncol(mat)} columns and
    \code{length(unique(by)) * n} rows, where \code{n} is the
    length of the result returned by \code{f}.
}
\description{
Apply a function columnwise to groups of rows
}

