% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaboliteqc.R
\name{copy_NAs}
\alias{copy_NAs}
\title{Copy NAs from one matrix to another}
\usage{
copy_NAs(from, to)
}
\arguments{
\item{from}{Matrix with NAs}

\item{to}{Matrix to which to copy NAs from \code{from}}
}
\value{
A matrix like \code{to} but with NAs where \code{from} has
    NAs.
}
\description{
Copy NAs from one matrix to another
}

