% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaboliteqc.R
\name{correct_for_factor}
\alias{correct_for_factor}
\title{Correct metabolite measurements for a factor}
\usage{
correct_for_factor(mat, x)
}
\arguments{
\item{mat}{Matrix with samples as columns and metabolites as rows}

\item{x}{Factor for which to correct the metabolite measurements in
\code{mat}}
}
\value{
A matrix of the same dimensions as \code{mat} where every
    metabolite's measurements were replaced by the residuals from
    regressing the metabolite on the \code{x} treated as a factor.
    If a metabolite consists only of NAs, the residuals for that
    metabolite will also be NAs.  The same happens when a
    metabolite has no valid variance (because there is only a
    single non-NA measurement) or when the metabolite has zero
    variance.
}
\description{
Replace the measurements of every metabolite by the residuals from
regressing the metabolite on \code{x}, where \code{x} is converted
to a factor.
}

