% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaboliteqc.R
\name{count_below_percentile}
\alias{count_below_percentile}
\title{Count number of groups with measurements below percentile per sample}
\usage{
count_below_percentile(mat, by, f, percentile, ...)
}
\arguments{
\item{mat}{Matrix with samples as columns and metabolites as rows}

\item{by}{Factor of length \code{nrow(mat)} used to group rows}

\item{f}{Function to "condense" the measurements of a sample for a
given group of rows to a single number
(\code{\link[base]{mean}}, \code{\link[stats]{median}},
\code{\link[base]{min}}, etc.)}

\item{percentile}{A number between 0 and 100.}

\item{\dots}{Passed on to \code{f}}
}
\value{
A integer vector with \code{colnames(mat)} as names
    containing the number of categories, that is the levels of
    \code{by}, for which a sample has measurements (as "condensed"
    by \code{f}) below \code{percentile}.  The percentiles are
    computed separately for every metabolite over the "condensed"
    metabolite measurements of all samples.
}
\description{
Count number of groups with measurements below percentile per sample
}

