% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaboliteqc.R
\name{draw_box}
\alias{draw_box}
\title{Draw box for boxplot}
\usage{
draw_box(stats, width, at)
}
\arguments{
\item{stats}{Numeric vector of length 5 containing the extreme of
the lower whisker, the lower `hinge', the media, the upper
`hinge' and the extreme of the upper whisker (see also
\code{\link[grDevices]{boxplot.stats}}).}

\item{width}{Width of the box}

\item{at}{x-coordinate of the (center of the) box}
}
\value{
Returns \code{NULL} invisibly.
}
\description{
Draw box for boxplot
}

