% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaboliteqc.R
\name{draw_outliers}
\alias{draw_outliers}
\title{Draw outliers for boxplot}
\usage{
draw_outliers(outliers, at)
}
\arguments{
\item{outliers}{y-coordinates of outliers}

\item{at}{Length-1 x-coordinate of outliers}
}
\value{
Returns \code{NULL} invisibly.
}
\description{
Draw outliers for boxplot
}

