% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaboliteqc.R
\name{draw_whiskers}
\alias{draw_whiskers}
\title{Draw whiskers for boxplot}
\usage{
draw_whiskers(stats, width, at)
}
\arguments{
\item{stats}{Numeric vector of length 5 containing the extreme of
the lower whisker, the lower `hinge', the media, the upper
`hinge' and the extreme of the upper whisker (see also
\code{\link[grDevices]{boxplot.stats}}).}

\item{width}{Width of horizontal endings of whiskers}

\item{at}{x-coordinate of whiskers}
}
\value{
Returns \code{NULL} invisibly.
}
\description{
Draw whiskers for boxplot
}

