% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaboliteqc.R
\name{each_metabolite}
\alias{each_metabolite}
\title{Apply function to every metabolite}
\usage{
each_metabolite(mat, f, ...)
}
\arguments{
\item{mat}{Matrix with samples as columns and metabolites as rows}

\item{f}{Function to apply to metabolites}

\item{\dots}{Further arguments passed to \code{f}}
}
\value{
If \code{f} returns a vector of length 1,
    \code{each_metabolite} returns a vector with one element per
    metabolite.  Otherwise if \code{f} returns a vector of the same
    length as its argument, \code{each_metabolite} returns a matrix
    with one row per metabolite, that is, the result will have the
    same dimensions as \code{mat}.
}
\description{
Apply function to every metabolite
}

