% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaboliteqc.R
\name{each_pathway}
\alias{each_pathway}
\title{Apply function to every pathway}
\usage{
each_pathway(mat, pathways, f, ...)
}
\arguments{
\item{mat}{Matrix with samples as columns and metabolites as rows}

\item{pathways}{Factor of length \code{nrow(mat)} used to split
metabolites by pathway}

\item{f}{Function to apply to pathway subsets of \code{mat}}

\item{\dots}{Further arguments passed to \code{f}}
}
\value{
A list with one element per pathway.  Elements are named
    using the corresponding level of \code{pathways}.
}
\description{
Apply function to every pathway
}

