% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaboliteqc.R
\name{each_run_day}
\alias{each_run_day}
\title{Apply function to every run day}
\usage{
each_run_day(mat, run_days, f, ...)
}
\arguments{
\item{mat}{Matrix with samples as columns and metabolites as rows}

\item{run_days}{Factor of length \code{ncol(mat)} used to split
samples by run day}

\item{f}{Function to apply to run day subsets of \code{mat}}

\item{\dots}{Further arguments passed to \code{f}}
}
\value{
A list with one element per run day.  Elements are named
    using the corresponding level of \code{run_days}.
}
\description{
Apply function to every run day
}

