% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaboliteqc.R
\name{each_sample}
\alias{each_sample}
\title{Apply function to every sample}
\usage{
each_sample(mat, f, ...)
}
\arguments{
\item{mat}{Matrix with samples as columns and metabolites as rows}

\item{f}{Function to apply to samples}

\item{\dots}{Further arguments passed to \code{f}}
}
\value{
If \code{f} returns a vector of length 1,
    \code{each_sample} returns a vector with one element per
    sample.  Otherwise if \code{f} returns a vector of the same
    length as its argument, \code{each_sample} returns a matrix
    with one column per sample, that is, the result will have the
    same dimensions as \code{mat}.
}
\description{
Apply function to every sample
}

