% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaboliteqc.R
\name{find_cell_indices}
\alias{find_cell_indices}
\title{Find indices of cells that satisfy a predicate}
\usage{
find_cell_indices(mat, margin, predicate, ...)
}
\arguments{
\item{mat}{Matrix}

\item{margin}{Whether to apply \code{predicate} to rows or to
columns (1 = rows, 2 = columns)}

\item{predicate}{A function that takes a vector (a row or a column
of \code{mat}) and returns a logical vector of the same length}

\item{\dots}{Passed on to \code{predicate}}
}
\value{
A two-column character matrix that can be used as index to
    \code{mat} to extract elements that satisfy \code{predicate}.
}
\description{
Find indices of cells that satisfy a predicate
}
\details{
Note that \code{predicate} can use information on all
    elements in a row (column) to decide whether a particular
    element of the row (column) satisfies the predicate or not.
}

