% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaboliteqc.R
\name{hist_NAs}
\alias{hist_NAs}
\title{Histogram of number of NAs}
\usage{
hist_NAs(mat, by, ..., main = NULL, xlab = NULL)
}
\arguments{
\item{mat}{Matrix}

\item{by}{An integer.  Whether to count NAs by rows or by columns
(1 = rows, 2 = columns).}

\item{main}{Title for plot}

\item{xlab}{Label for x-axis}

\item{\dots}{Arguments passed on to \code{\link[graphics]{hist}}}
}
\value{
Returns \code{NULL} invisibly.
}
\description{
Histogram of number of NAs
}

