% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaboliteqc.R
\name{mark_outliers}
\alias{mark_outliers}
\title{Mark outliers}
\usage{
mark_outliers(x)
}
\arguments{
\item{x}{A vector of numbers}
}
\value{
A logical vector of the same length as \code{x} where
    elements are \code{TRUE} if the corresponding element of
    \code{x} is an outlier among the the values in \code{x}, and
    otherwise \code{FALSE}.
}
\description{
Mark outliers
}

