% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaboliteqc.R
\name{plot_samples_with_low_metabolites}
\alias{plot_samples_with_low_metabolites}
\title{Show how many samples have low measurements in how many groups}
\usage{
plot_samples_with_low_metabolites(mat, by, percentiles, by_label = NULL,
  xlim = NULL, ylim = NULL, main = NULL)
}
\arguments{
\item{mat}{Matrix with samples as columns and metabolites as rows}

\item{by}{Factor of length \code{nrow(mat)} used to group rows}

\item{percentiles}{Integer vector with numbers between 0 and 100.}

\item{by_label}{Label used in plot annotations to refer to the
groups defined by \code{by}}

\item{xlim}{Limits plotting range on x-axis}

\item{ylim}{Limits plotting range on y-axis}

\item{main}{Title}
}
\value{
Returns \code{NULL} invisibly.
}
\description{
Plot the number of samples for which a given number of groups has
mean metabolite measurements below a given percentile.
}
\details{
Algorithm: For every sample and every group of rows the
    mean metabolite measurement is computed.  This results in every
    sample having one value per group (the mean metabolite
    measurement).  We then count for every sample the number of
    groups in which the sample's mean metabolite measurement is
    less that or equal to the given percentile of the distribution
    of all mean metabolite measurements in that group.  Finally we
    plot the number of groups N against the number of samples with
    mean metabolite measurements less than or equal to the given
    percentile in N groups.
}

