% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaboliteqc.R
\name{replace_NAs_with_column_means}
\alias{replace_NAs_with_column_means}
\alias{replace_NAs_with_row_means}
\title{Replace missing values with column/row means}
\usage{
replace_NAs_with_column_means(mat)

replace_NAs_with_row_means(mat)
}
\arguments{
\item{mat}{Matrix}
}
\value{
A matrix just like \code{mat} but with NAs in column/row k
    replaced by the mean over column/row k.
}
\description{
Replace missing values with column/row means
}

