% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaboliteqc.R
\name{same}
\alias{same}
\title{Check whether objects are identical}
\usage{
same(..., key = NULL)
}
\arguments{
\item{key}{Function to apply to objects before comparison (see
Details)}

\item{\dots}{Any number of objects}
}
\value{
Returns \code{TRUE} if all objects are identical, otherwise
    returns \code{FALSE}.
}
\description{
Check whether objects are identical
}
\details{
If \code{key} is \code{NULL}, objects are compared
    directly.  If \code{key} is a function, it will be applied to
    objects and the return values will be compared.  This makes it
    possible to compare objects in terms of arbitrary attributes.
}
\examples{
\dontrun{
# Do data frames d1, d2, and d3 have the same row names?
same(d1, d2, d3, key = rownames)
}
}

