% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaboliteqc.R
\name{select_if}
\alias{select_if}
\title{Select rows/columns of a matrix using a predicate}
\usage{
select_if(predicate, mat, margin, ...)
}
\arguments{
\item{predicate}{A function taking a row/column as first argument
and returning \code{TRUE} or \code{FALSE}}

\item{mat}{Matrix}

\item{margin}{Whether to select rows or columns (1 = rows, 2 =
columns).}

\item{\dots}{Passed on to \code{predicate}}
}
\value{
If \code{margin} is 1, a matrix with the same number of
    columns as \code{mat} and including all rows of \code{mat} for
    which \code{predicate} returned \code{TRUE}.  If \code{margin}
    is 2, a matrix with the same number of rows as \code{mat} and
    including all columns of \code{mat} for which \code{predicate}
    returned \code{TRUE}.
}
\description{
Select rows/columns of a matrix using a predicate
}

