% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaboliteqc.R
\name{summarize_NAs}
\alias{summarize_NAs}
\title{Find number of missing values and missingness rate}
\usage{
summarize_NAs(mat, by)
}
\arguments{
\item{mat}{Matrix}

\item{by}{An integer.  Dimension for which to compute missingness
statistics (1 = rows, 2 = columns).  Same as \code{MARGIN}
argument to \code{\link[base]{apply}}.}
}
\value{
A data frame with columns "id", "missings", and
    "missingness_rate" containing row (column) names, number of
    missings per row (column), and missingness rate per row
    (column) if \code{by} is 1 (2).
}
\description{
Find number of missing values and missingness rate
}

