% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaboliteqc.R
\name{write.delim}
\alias{write.delim}
\title{Save dataset as tab-delimited plain text file}
\usage{
write.delim(x, file, ...)
}
\arguments{
\item{x}{Dataset}

\item{file}{Filename}

\item{\dots}{Passed on to \code{\link[utils]{write.table}}}
}
\value{
Returns \code{NULL} invisibly.
}
\description{
Save dataset as tab-delimited plain text file
}
\details{
This function is a wrapper around
    \code{\link[utils]{write.table}}.  Refer to the latter's
    documentation for more details.
}

