% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaboliteqc.R
\name{draw_boxplot}
\alias{draw_boxplot}
\title{Draw boxplot}
\usage{
draw_boxplot(stats, at)
}
\arguments{
\item{stats}{Output of \code{\link[grDevices]{boxplot.stats}}}

\item{at}{x-coordinate of the boxplot}
}
\value{
Returns \code{NULL} invisibly.
}
\description{
Draw boxplot
}

